Adjust python for new API
