+++ python/dbsp/__init__.py
@@ -1,6 +1,5 @@
-from .project import DBSPProject
-from .pipeline import DBSPPipeline
-from .config import DBSPPipelineConfig
+from .program import DBSPProgram
+from .pipeline import DBSPPipelineConfig
 from .error import DBSPServerError
 from .error import TimeoutException
 from .error import CompilationException
